+++ ktfmt_idea_plugin/src/main/java/com/facebook/ktfmt/intellij/UiFormatterStyle.java
@@ -16,7 +16,6 @@
 
 package com.facebook.ktfmt.intellij;
 
-import static com.facebook.ktfmt.format.Formatter.DROPBOX_FORMAT;
 import static com.facebook.ktfmt.format.Formatter.GOOGLE_FORMAT;
 import static com.facebook.ktfmt.format.Formatter.KOTLINLANG_FORMAT;
 
@@ -25,7 +24,6 @@ import com.facebook.ktfmt.format.FormattingOptions;
 /** Configuration options for the formatting style. */
 enum UiFormatterStyle {
   DEFAULT("Default", new FormattingOptions()),
-  DROPBOX("Dropbox", DROPBOX_FORMAT),
   GOOGLE("Google (internal)", GOOGLE_FORMAT),
   KOTLINLANG("Kotlinlang", KOTLINLANG_FORMAT);